working command parsing from uart. set_target working from uart

+++ main/cmd.h
@@ -18,7 +18,7 @@ typedef enum {
 } COMM_TYPE;
 
 
-typedef bool (*CMD_RESPONSE_CB)(int socket, COMM_TYPE Message_Type, void* i_pBuf, uint8_t size);
+typedef bool (*CMD_RESPONSE_CB)(int socket, COMM_TYPE Message_Type, void* i_pBuf, uint16_t size);
 
 typedef struct {
 	CMD_RESPONSE_CB	p_cbSend;
